+++ index.js
@@ -2,6 +2,7 @@
 module.exports = function (expected) {
 	var steps = expected || [];
 	var step = 0;
+	var done = [];
 
 	var stub = function () {
 		if (step > steps.length - 1) {
@@ -13,16 +14,24 @@ module.exports = function (expected) {
 			throw new Error(tooManyTimes);
 		}
 
-		return steps[step++].apply(null, arguments);
+		var result = steps[step++].apply(null, arguments);
+
+		if (step == steps.length) done.forEach((fn) => fn());
+
+		return result;
 	};
 
 	stub.add = function (fn, count) {
-	  count = count || 1;
-	  for(var i = 0; i < count; i++) {
-	    steps.push(fn);
-	  }
+		count = count || 1;
+		for(var i = 0; i < count; i++) {
+			steps.push(fn);
+		}
 		return stub;
 	};
+
+	stub.onDone = function (fn) {
+		done.push(fn);
+	}
 
 	stub.onCall = function (i, fn) {
 		steps[i] = fn;
@@ -45,8 +54,10 @@ module.exports = function (expected) {
 		step = 0;
 		if (Array.isArray(newsteps)) {
 			steps = newsteps;
+			done = [];
 		} else if (newsteps === true) {
 			steps = [];
+			done = [];
 		}
 		return stub;
 	};
@@ -58,6 +69,10 @@ module.exports = function (expected) {
 	stub.getStep = function () {
 		return step;
 	};
+
+	stub.getCount = function () {
+		return steps.length;
+	}
 
 	return stub;
 };